longitude fixed

+++ models/records.go
@@ -8,8 +8,8 @@ type Record struct {
 	Time      time.Time
 	Angle     int16
 	Speed     int16
-	Longitude int32
-	Latitude  int32
+	Longitude float32
+	Latitude  float32
 }
 type Location struct {
 	Type        string
